Ajustando services e dtos

+++ barbeariaPro/barbeariaPro/Services/ClienteService.cs
@@ -1,15 +1,14 @@
 ﻿using Microsoft.EntityFrameworkCore;
-
-namespace barbeariaPro.Services;
-
 using barbeariaPro.Models;
 using barbeariaPro.dbContext;
+
+namespace barbeariaPro.Services;
 
 public class ClienteService
 {
     private readonly AppDbContext _context;
 
-    public ClienteService(AppDbContext context) 
+    public ClienteService(AppDbContext context)
     {
         _context = context;
     }
@@ -21,24 +20,42 @@ public class ClienteService
 
     public async Task<Cliente?> ObterClientePorId(int id)
     {
-        return await _context.Clientes.FindAsync(id);
+        return await _context.Clientes
+                             .FirstOrDefaultAsync(c => c.Id == id); 
     }
 
     public async Task<Cliente> AdicionarCliente(Cliente novoCliente)
     {
+        if (await CpfExiste(novoCliente.Cpf))
+        {
+            throw new Exception("Já existe um cliente com esse CPF.");
+        }
+
         _context.Clientes.Add(novoCliente);
         await _context.SaveChangesAsync();
         return novoCliente;
     }
-    
+
     public async Task AtualizarCliente(Cliente cliente)
     {
+        var clienteExistente = await ObterClientePorId(cliente.Id);
+        if (clienteExistente == null)
+        {
+            throw new Exception("Cliente não encontrado.");
+        }
+
         _context.Clientes.Update(cliente);
         await _context.SaveChangesAsync();
     }
 
     public async Task DeletarCliente(Cliente cliente)
     {
+        var clienteExistente = await ObterClientePorId(cliente.Id);
+        if (clienteExistente == null)
+        {
+            throw new Exception("Cliente não encontrado.");
+        }
+
         _context.Clientes.Remove(cliente);
         await _context.SaveChangesAsync();
     }
